changed some variable types

+++ src/Model/Producte.java
@@ -4,17 +4,17 @@ import java.util.Objects;
 
 abstract class Producte {
     String nom;
-    int preu;
+    float preu;
     String codi_barres;
 
-    public Producte(String nom, int preu, String codi_barres) {
+    public Producte(String nom, float preu, String codi_barres) {
         this.nom = nom;
         this.preu = preu;
         this.codi_barres = codi_barres;
     }
 
     public String getCodi_barres() {return codi_barres;}
-    public int getPreu() {return preu;}
+    public float getPreu() {return preu;}
     public String getNom() {return nom;}
 
     public abstract double calcularPreu();
